타입 변경 및 클래스 명 변경

+++ XmlParseLib/XmlNode.h
@@ -78,7 +78,7 @@ inline XmlData& XmlNode::GetData()
 }
 inline XmlNode::~XmlNode()
 {
-	for (int i = 0; i < _child.size(); i++)
+	for (size_t i = 0; i < _child.size(); i++)
 	{
 		delete _child[i];
 		_child[i] = NULL;
